requirements.txt
google-cloud-storage
google-cloud-firestore
grpcio 
flask
flask-cors
